news_agent/requirements.txt
google-adk
Deprecated
newsapi-python
newspaper3k~=0.2.8
python-dotenv~=1.1.0
lxml[html_clean]
protobuf~=5.29.4